requirements.txt
asgiref==3.2.10
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
Django==3.1.1
django-debug-toolbar==2.2
django-extra-fields==3.0.1
djangorestframework==3.11.1
dnspython==2.0.0
drf-base64==2.0
exchangelib==3.2.1
idna==2.10
isodate==0.6.0
lxml==4.5.2
ntlm-auth==1.5.0
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.10
